[P24-77] feat: 이사 플랜 삭제 버튼 클릭 시 삭제 확인 기능 추가
- 기본 브라우저 confirm 창으로 구현
- 추후 여건이 된다면 커스텀 모달 구현해서 수정

+++ FE/src/pages/PlanSetting.jsx
@@ -64,11 +64,17 @@ export default function PlanSetting() {
   };
 
   // 이사 플랜 삭제
-  const deletePlan = async (e) => {
-    try {
-      await planApi.deletePlan(movingPlanId);
-      navigate('/plans');
-    } catch (error) {}
+  const deletePlan = async () => {
+    const confirmDelete = window.confirm(
+      '이 이사 플랜과 관련된 모든 데이터가 삭제됩니다.\n정말 삭제하시겠습니까?',
+    );
+
+    if (confirmDelete) {
+      try {
+        await planApi.deletePlan(movingPlanId);
+        navigate('/plans');
+      } catch (error) {}
+    }
   };
 
   // CSS
